+++ CHANGES.rst
@@ -5,6 +5,7 @@ tip (unreleased)
 ----------------
 - Do NOT delete the history elements when a user is deleted.
 - Add support for ``latest``
+- Allow setting a reason for change. [using option changeReason]
 
 1.5.3 (2014-11-18)
 ------------------